+++ database/migrations/2022_08_18_094017_create_transactions_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->foreignId('account_id')->nullable()->constrained('accounts')->onDelete('cascade');
             $table->string("description");
             $table->string("type");
-            $table->integer("amount");
+            $table->decimal("amount",8,2);
             $table->timestamps();
             $table->softDeletes();
         });
